Fix component.get_or_create function
the get_or_create function is using the same
parameters for identifying a component and
for creating the component if it does not
exists. This might lead to not getting the
components because of the extras filters that
may be the same in two components.

This patch ensure that a component exists or
not by using only the filed that identify
uniquely a component.

Change-Id: I324ceba0e3a22b62d79d369eb1f5105b297d9fbd
Depends-On: https://softwarefactory-project.io/r/c/dci-packaging/+/23842